.......PS. [ZBX-22336] added support of log level control for threaded components
Merge in ZBX/zabbix from feature/ZBX-22336-6.3 to release/6.4

* commit '277b2a01a16151d3a50546eb2137a45a7dd429dd':
  .D........ [ZBX-22336] added changelog entry
  .......PS. [ZBX-22336] added SIGINT blockign in preprocessing worker threads
  .......PS. [ZBX-22336] block SIGUSR1 in threads - only parent process should be receiving rtc signals
  .......... [ZBX-22336] moved component related log functions from windows code
  .......PS. [ZBX-22336] fixed inability to set log level for last worker process
  .......PS. [ZBX-22336] fixed build warnings wihtout net-snmp
  ...G...PS. [ZBX-22336] fixed warnings
  ...G...PS. [ZBX-22336] fixed log level handling in non-component threads
  .......... [ZBX-22336] style fix
  .......PS. [ZBX-22336] fixed warnings
  .......PS. [ZBX-22336] fixed component log level limit checks
  .......PS. [ZBX-22336] removed redundant code
  .......PS. [ZBX-22336] added support for preprocessing worker loglevel control
  .......PS. [ZBX-22336] fixed subscription for all processes of one type, minor refactoring
  .......PS. [ZBX-22336] added signal blocking in preprocessing worker threads
  .......PS. [ZBX-22336] added RTC subcription support for services
  ........S. [ZBX-22336] fixed uninitialized connector statistics
  .......PS. [ZBX-22336] fixed runtime option handling
  .......PS. [ZBX-22336] refactored loglevel/profiler rtc message dispatching
  .......PS. [ZBX-22336] refactored rtc option parsing

(cherry picked from commit 795a510f02d2d5f70631dde1f33c42622c27b053)

+++ src/libs/zbxrtc/rtc.c
@@ -92,11 +92,11 @@ static int	rtc_option_get_ui64(const char *param, size_t *size, zbx_uint64_t *va
 
 	if (ptr != param)
 	{
-		if (FAIL == zbx_is_uint64_n(param, ptr - param, value))
+		if (FAIL == zbx_is_uint64_n(param, (size_t)(ptr - param), value))
 			return FAIL;
 	}
 
-	*size = ptr - param;
+	*size = (size_t)(ptr - param);
 
 	return SUCCEED;
 }
@@ -150,7 +150,7 @@ int	rtc_option_get_process_type(const char *param, size_t *size, int *proc_type,
 	int		ret = FAIL;
 
 	if (NULL != (ptr = strchr(param, ',')))
-		*size = ptr - param;
+		*size = (size_t)(ptr - param);
 	else
 		*size = strlen(param);
